requirements.txt
ultralytics
pyttsx3
SpeechRecognition
pyaudio
opencv-python
Pillow
tkinter
